Решили задачу

+++ Ex002/Library1.cs
@@ -2,9 +2,9 @@ using System;
 
 public static class Library1
 {
-    public static double GetNumber(string text)
+    public static int GetNumber(string text)
     {
         Console.Write(text);
-        return Convert.ToDouble(Console.ReadLine());
+        return Convert.ToInt32(Console.ReadLine());
     }
 }
